funktionen in parse_affiliations.php ausgelagert
save_authors nutzt funktionen, Speicherung funktioniert

save_contact_persons funktioniert ebenfalls

TODO:
- save_contributors
- export freestyle xml testen+anpassen
